+++ src/app/(core)/programmes/[subjectPhaseSlug]/[tab]/Components/ProgrammeView.test.tsx
@@ -31,8 +31,6 @@ beforeEach(() => {
 });
 
 const defaultProps = {
-  subjectPhaseSlug: "science-secondary-aqa",
-  subjectTitle: "Science",
   curriculumSelectionSlugs: {
     phaseSlug: "secondary",
     subjectSlug: "science",
@@ -43,6 +41,7 @@ const defaultProps = {
     phaseTitle: "Secondary",
     examboardTitle: "AQA",
   },
+  subjectPhaseSlug: "science-secondary-aqa",
   ks4Options: [],
   curriculumUnitsFormattedData: formatCurriculumUnitsData(
     curriculumUnitsTabFixture(),
@@ -79,7 +78,7 @@ describe("ProgrammeView", () => {
   });
   it("renders the correct tab content for units", () => {
     render(<ProgrammeView {...defaultProps} />);
-    const heading = screen.getByRole("heading", { name: "Year 7" });
+    const heading = screen.getByRole("heading", { name: "Year 7 units" });
     expect(heading).toBeInTheDocument();
   });
   it("renders the correct tab content for overview", () => {