+++ src/app/(core)/programmes/[subjectPhaseSlug]/[tab]/Components/UnitSequence/UnitSequenceView.tsx
@@ -1,14 +1,16 @@
-"use client";
-import { OakBox, OakHeading } from "@oaknational/oak-components";
+import {
+  OakBox,
+  OakGrid,
+  OakGridArea,
+  OakHeading,
+} from "@oaknational/oak-components";
 
 import ProgrammePageFiltersDesktop from "../Filters/ProgrammePageFiltersDesktop";
 import ProgrammePageFiltersMobile from "../Filters/ProgrammePageFiltersMobile";
 
 import ProgrammeSequence from "./Sequence";
 
-import { CurricVisualiserLayout } from "@/components/CurriculumComponents/CurricVisualiserLayout";
 import ScreenReaderOnly from "@/components/SharedComponents/ScreenReaderOnly";
-import useMediaQuery from "@/hooks/useMediaQuery";
 import useAnalytics from "@/context/Analytics/useAnalytics";
 import useAnalyticsPageProps from "@/hooks/useAnalyticsPageProps";
 import {
@@ -21,17 +23,13 @@ import {
   highlightedUnitCount,
 } from "@/utils/curriculum/filteringApp";
 import { CurriculumFilters } from "@/utils/curriculum/types";
-import {
-  CurriculumSelectionSlugs,
-  CurriculumSelectionTitles,
-} from "@/utils/curriculum/slugs";
+import { CurriculumSelectionSlugs } from "@/utils/curriculum/slugs";
 import { Ks4Option } from "@/node-lib/curriculum-api-2023/queries/curriculumPhaseOptions/curriculumPhaseOptions.schema";
 
 export type UnitSequenceViewProps = {
   filters: CurriculumFilters;
   setFilters: (newFilters: CurriculumFilters) => void;
   curriculumSelectionSlugs: CurriculumSelectionSlugs;
-  curriculumSelectionTitles: CurriculumSelectionTitles;
   curriculumUnitsFormattedData: CurriculumUnitsFormattedData;
   ks4Options: Ks4Option[];
   trackingData: CurriculumUnitsTrackingData;
@@ -41,15 +39,12 @@ export const UnitSequenceView = ({
   filters,
   setFilters,
   curriculumSelectionSlugs,
-  curriculumSelectionTitles,
   curriculumUnitsFormattedData,
   ks4Options,
   trackingData,
 }: UnitSequenceViewProps) => {
-  const isDesktop = useMediaQuery("desktop");
   const { yearData, threadOptions } = curriculumUnitsFormattedData;
   const { ks4OptionSlug } = curriculumSelectionSlugs;
-  const { subjectTitle } = curriculumSelectionTitles;
 
   const unitCount = getNumberOfSelectedUnits(yearData, filters);
 
@@ -75,38 +70,43 @@ export const UnitSequenceView = ({
   };
 
   return (
-    <OakBox
-      id="programme-units"
-      aria-labelledby="programme-unit-sequence-heading"
-      tabIndex={-1}
-      $mh={"auto"}
-      $mt={["spacing-0", "spacing-48", "spacing-48"]}
-      $width={"100%"}
-      as="section"
-    >
-      <ScreenReaderOnly>
-        <OakHeading
-          id="programme-unit-sequence-heading"
-          tag="h2"
-          $mb="spacing-24"
-          $ml={["spacing-16", "spacing-0"]}
-          $font={["heading-5", "heading-4"]}
-        >
-          Unit sequence
-        </OakHeading>
-      </ScreenReaderOnly>
-      {!isDesktop && (
-        <ProgrammePageFiltersMobile
-          filters={filters}
-          onChangeFilters={onChangeFilters}
-          data={curriculumUnitsFormattedData}
-          slugs={curriculumSelectionSlugs}
-          ks4Options={ks4Options}
-        />
-      )}
-      <CurricVisualiserLayout
-        filters={
-          isDesktop ? (
+    <OakBox $ph={["spacing-20", "spacing-40"]}>
+      <OakBox
+        id="programme-units"
+        aria-labelledby="programme-unit-sequence-heading"
+        tabIndex={-1}
+        $mh={"auto"}
+        $mt={["spacing-0", "spacing-48", "spacing-48"]}
+        $width={"100%"}
+        $maxWidth={"spacing-1280"}
+        $color="text-primary"
+        as="section"
+      >
+        <ScreenReaderOnly>
+          <OakHeading
+            id="programme-unit-sequence-heading"
+            tag="h2"
+            $mb="spacing-24"
+            $ml={["spacing-16", "spacing-0"]}
+            $font={["heading-5", "heading-4"]}
+          >
+            Unit sequence
+          </OakHeading>
+        </ScreenReaderOnly>
+        <OakBox $display={["block", "block", "none"]}>
+          <ProgrammePageFiltersMobile
+            filters={filters}
+            onChangeFilters={onChangeFilters}
+            data={curriculumUnitsFormattedData}
+            slugs={curriculumSelectionSlugs}
+            ks4Options={ks4Options}
+          />
+        </OakBox>
+        <OakGrid $cg={"spacing-16"}>
+          <OakGridArea
+            $colSpan={[12, 12, 3]}
+            $display={["none", "none", "block"]}
+          >
             <ProgrammePageFiltersDesktop
               filters={filters}
               onChangeFilters={onChangeFilters}
@@ -114,32 +114,30 @@ export const UnitSequenceView = ({
               slugs={curriculumSelectionSlugs}
               ks4Options={ks4Options}
             />
-          ) : null
-        }
-        units={
-          <ProgrammeSequence
-            filters={filters}
-            ks4OptionSlug={ks4OptionSlug}
-            ks4Options={ks4Options}
-            yearData={yearData}
-            threadOptions={threadOptions}
-          />
-        }
-        curriculumSeoText={undefined} // TD: [integrated journey] seo text
-        subjectTitle={subjectTitle}
-      />
-      <ScreenReaderOnly aria-live="polite" aria-atomic="true">
-        <p>
-          {unitCount} {unitCount === 1 ? "unit" : "units"} shown,
-        </p>
-        {filters.threads[0] && (
+          </OakGridArea>
+          <OakGridArea $colSpan={[12, 12, 9]}>
+            <ProgrammeSequence
+              filters={filters}
+              ks4OptionSlug={ks4OptionSlug}
+              ks4Options={ks4Options}
+              yearData={yearData}
+              threadOptions={threadOptions}
+            />
+          </OakGridArea>
+        </OakGrid>
+        <ScreenReaderOnly aria-live="polite" aria-atomic="true">
           <p>
-            {highlightedUnits}
-            {highlightedUnits === 1 ? "unit" : "units"}
-            highlighted
+            {unitCount} {unitCount === 1 ? "unit" : "units"} shown,
           </p>
-        )}
-      </ScreenReaderOnly>
+          {filters.threads[0] && (
+            <p>
+              {highlightedUnits}
+              {highlightedUnits === 1 ? "unit" : "units"}
+              highlighted
+            </p>
+          )}
+        </ScreenReaderOnly>
+      </OakBox>
     </OakBox>
   );
 };